APPS * Fixing bug where checksum calculation was throwing NPE when using relay transfers due to a bug in the URLCopy code deleting the temp directory holding the intermediate file. The problem was that when the source system was type LOCAL, the intermediate file was the source file, so the file was being deleted. Added a check for the source type to avoid this situation.
SYSTEMS
* Adding a null check to children of recursive transfers. This was causing failures in transfers when the destination was not recognized,

+++ agave-systems/systems-core/src/main/java/org/iplantc/service/transfer/URLCopy.java
@@ -290,7 +290,8 @@ public class URLCopy
     						TransferTaskDao.persist(transferTask);
     					}
     					
-    					if (childTransferTask.getStatus().equals(TransferStatusType.CANCELLED)) {
+    					if (childTransferTask != null && 
+    							TransferStatusType.CANCELLED == childTransferTask.getStatus()) {
     						transferTask.setStatus(TransferStatusType.CANCELLED);
     						break;
     					}
@@ -642,8 +643,12 @@ public class URLCopy
 				        getProtocolForClass(sourceClient.getClass()),
 				        getProtocolForClass(destClient.getClass())));
 			}
-			log.info("Deleting relay cache file " + tempDir.getAbsolutePath());
-			FileUtils.deleteQuietly(tempDir);
+			if (sourceClient instanceof Local) {
+				FileUtils.deleteQuietly(tempDir);
+			}
+			else {
+				log.info("Deleting relay cache file " + tempDir.getAbsolutePath());
+			}
 		}
 		
 	}
